feat: routeCreate 不覆盖已创建路由文件，功能支持

+++ tools/routeCreate.js
@@ -1,3 +1,4 @@
+// 只对pages下的文件夹下的.vue文件，生成路由。直接位于pages下的.vue，由开发者手动在router/index.js配置
 const $inquirer = require('inquirer')
 const $fse = require('fs-extra')
 const $chalk = require('chalk')
@@ -7,65 +8,16 @@ function isFolder(path) {
 	return stat.isDirectory()
 }
 
-function createFile(items, rewrite) {
-	items.forEach(item => {
-		// 处理直接存放于 pages 下的 .vue 文件
-		if (Array.isArray(item)) {
-			$fse.pathExists('src/router/default.js', (err, exists) => {
-				if (exists) {
-					if (rewrite) {
-						createDefaultFile(item)
-					}
-				} else {
-						createDefaultFile(item)
-				}
-			})
-		} else {
-			// 处理一般情况（页面存放于 pages 下的 自定义目录下）
-			$fse.pathExists(`src/router/${item}.js`, (err, exists) => {
-				if (exists) {
-					if (rewrite) {
-						createFileContent('src/pages', item)
-					}
-				} else { // 一般情况，直接创建新的路由文件
-					createFileContent('src/pages', item)
-				}
-			})
-		}
-	})
-}
-
-// default.js 创建
-function createDefaultFile(fileNames) {
-	let routerList = ''
-	let imports = ''
-	fileNames.forEach(fileName => {
-		// fileName 截掉.vue
-		const name = `pages_${fileName.substring(0, fileName.lastIndexOf('.'))}`
-		// 构造 import 引入语句
-		const importStr = `const ${name} = () => import('pages/${fileName}')\n`
-		// 路由默认使用 pages 下的文件夹目录结构（将 path pages/ 及之前的路径截掉。将.vue字符截掉）
-		const routerPath = fileName.substring(0, fileName.lastIndexOf('.'))
-		// 构造路由描述字符串
-		const routerStr = `{\n\tpath: '/${routerPath}',\n\tcomponent: ${name}\n},\n`
-		// import 语句生成
-		imports = imports + importStr
-		// routerList 生成
-		routerList = routerList + routerStr
-	})
-	// 截掉 routerList 最后一个逗号 + 最后一个换行符
-	routerList = routerList.substring(0, routerList.length - 2)
-	// 路由文件内容生成
-	const contentStr = `${imports}\nconst routerList = [${routerList}]\n\nexport default routerList\n`
-	// 执行文件创建与输出
-	outputFile('default', contentStr)
-}
-
 // 生成指定路由文件（${folderName}.js）的内容 => pages下的 folderName 对应 router 下生成的 fileName
-function createFileContent(rootPath, folderName) {
+function createFile(rootPath, folderName, overwrite) {
 	const arr = []
 	let routerList = ''
 	let imports = ''
+	// 已存在路由文件的内容字符串
+	let fileStr = ''
+	if (!overwrite) {
+		fileStr = $fse.readFileSync(`src/router/${folderName}.js`).toString()
+	}
 	// 拿到当前文件夹下，所有文件的路径
 	getPagePaths(rootPath, folderName, arr)
 	arr.forEach(path => {
@@ -87,18 +39,34 @@ function createFileContent(rootPath, folderName) {
 			routerPath = `/${folderName}`
 		}
 		// 构造路由描述字符串
-		const routerStr = `{\n\tpath: '${routerPath}',\n\tcomponent: ${name}\n},\n`
-		// import 语句生成
-		imports = imports + importStr
-		// routerList 生成
-		routerList = routerList + routerStr
+		const routerStr = `\t{\n\t\tpath: '${routerPath}',\n\t\tcomponent: ${name}\n\t},\n`
+		if (!overwrite) {
+			if (fileStr.indexOf(importStr) === -1) {
+				// 在原有路由文件头部，增加import语句
+				fileStr = importStr + fileStr
+				const strArr = fileStr.split('[')
+				// 截掉routerStr最后一个换行
+				strArr[0] = (strArr[0] + '[\n' + routerStr).slice(0, -1)
+				fileStr = strArr.join('')
+			}
+		} else {
+			// import 语句生成
+			imports = imports + importStr
+			// routerList 生成
+			routerList = routerList + routerStr
+		}
 	})
-	// 截掉 routerList 最后一个逗号 + 最后一个换行符
-	routerList = routerList.substring(0, routerList.length - 2)
-	// 路由文件内容生成
-	const contentStr = `${imports}\nconst routerList = [${routerList}]\n\nexport default routerList\n`
-	// 执行文件创建与输出
-	outputFile(folderName, contentStr)
+	if (!overwrite) {
+		// 执行文件创建与输出
+		outputFile(folderName, fileStr)
+	} else {
+		// 截掉 routerList 最后一个逗号 + 最后一个换行符
+		routerList = routerList.slice(0, -2)
+		// 路由文件内容生成
+		const contentStr = `${imports}\nconst routerList = [\n${routerList}\n]\n\nexport default routerList\n`
+		// 执行文件创建与输出
+		outputFile(folderName, contentStr)
+	}
 }
 
 // 获取所有的 vue page 文件的路径
@@ -107,7 +75,7 @@ function getPagePaths(rootPath, folderName, arr) {
 	data && data.forEach(item => {
 		// 如果存在二级、三级...目录
 		if (isFolder(`${rootPath}/${folderName}/${item}`)) {
-			getPagePaths(`${rootPath}/${folderName}`, item)
+			getPagePaths(`${rootPath}/${folderName}`, item, arr)
 		} else {
 			arr.push(`${rootPath}/${folderName}/${item}`)
 		}
@@ -118,7 +86,7 @@ function getPagePaths(rootPath, folderName, arr) {
 function outputFile(fileName, content) {
 	$fse.outputFile(`src/router/${fileName}.js`, content, function(err) {
 		if (err) {
-			return
+			console.log($chalk.red('\n啊哦~文件生成出错 T_T\n'))
 		}
 	})
 }
@@ -128,33 +96,36 @@ $inquirer.prompt([
 	{
 		type: 'confirm',
 		name: 'rewrite',
-		message: $chalk.yellow('我们会根据pages的目录结构，全自动帮您生成router配置。请先确认：对于已存在的路由文件，是否执行覆盖？'),
+		message: '我们会根据pages的目录结构，全自动帮您生成router配置。请选择：对于已存在的路由文件，是否执行覆盖？',
 		default: true
 	},
 	{
 		type: 'confirm',
 		name: 'confirm',
-		message: $chalk.yellow('即将执行路由生成，请确认'),
+		message: '即将执行路由生成，请确认',
 		default: true
 	}
 ]).then(answer => {
 	if (answer.confirm) {
 		$fse.readdir('src/pages', 'binary', function(err, data) {
-			const folders = []
-			const defaults = []
 			data && data.forEach(pageItem => {
-				// 遍历pages下的每一个文件/文件夹
+				// 遍历pages下的每一个文件夹
 				if (isFolder(`src/pages/${pageItem}`)) {
-					folders.push(pageItem)
-				} else {
-					defaults.push(pageItem)
+					// 判断当前文件夹对应路由文件是否存在
+					$fse.pathExists(`src/router/${pageItem}.js`, (err, exists) => {
+						// 如果路由文件已存在
+						if (exists) {
+								createFile('src/pages', pageItem, answer.rewrite)
+						} else {
+							// 不存在，直接创建新的路由文件
+							createFile('src/pages', pageItem, true)
+						}
+					})
 				}
 			})
-			if (defaults.length > 0) {
-				folders.push(defaults)
-			}
-			createFile(folders, answer.rewrite)
 		})
+	} else {
+		console.log($chalk.yellow('\n结束路由创建\n'))
 	}
 })
